[BEAM-6431] Move ExecutionStateSampler and ExecutionStateTracker into runners-core-java
