+++ runners/core-java/src/main/java/org/apache/beam/runners/core/metrics/ExecutionStateSampler.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.runners.dataflow.worker.util.common.worker;
+package org.apache.beam.runners.core.metrics;
 
 import static org.apache.beam.vendor.guava.v20_0.com.google.common.base.Preconditions.checkNotNull;
 
@@ -28,6 +28,7 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
+import javax.annotation.Nullable;
 import org.apache.beam.vendor.guava.v20_0.com.google.common.annotations.VisibleForTesting;
 import org.apache.beam.vendor.guava.v20_0.com.google.common.util.concurrent.ThreadFactoryBuilder;
 import org.joda.time.DateTimeUtils.MillisProvider;
@@ -66,7 +67,7 @@ public class ExecutionStateSampler {
 
   private static final long PERIOD_MS = 200;
 
-  private Future<Void> executionSamplerFuture = null;
+  @Nullable private Future<Void> executionSamplerFuture = null;
 
   /**
    * Called to start the ExecutionStateSampler. Until the returned {@link Closeable} is closed, the
@@ -141,7 +142,7 @@ public class ExecutionStateSampler {
   }
 
   /**
-   * Deregister tracker after MapTask completes
+   * Deregister tracker after MapTask completes.
    *
    * <p>This method needs to be synchronized to prevent race condition with sampling thread
    */
@@ -161,7 +162,7 @@ public class ExecutionStateSampler {
   }
 
   /**
-   * Attributing sampling time to trackers
+   * Attributing sampling time to trackers.
    *
    * <p>This method needs to be synchronized to prevent race condition with removing tracker
    */
