+++ runners/google-cloud-dataflow-java/worker/src/main/java/org/apache/beam/runners/dataflow/worker/ContextActivationObserver.java
@@ -20,7 +20,7 @@ package org.apache.beam.runners.dataflow.worker;
 import com.google.auto.service.AutoService;
 import java.io.Closeable;
 import java.util.ServiceLoader;
-import org.apache.beam.runners.dataflow.worker.util.common.worker.ExecutionStateTracker;
+import org.apache.beam.runners.core.metrics.ExecutionStateTracker;
 import org.apache.beam.sdk.annotations.Experimental;
 
 /**
