+++ runners/google-cloud-dataflow-java/worker/src/main/java/org/apache/beam/runners/dataflow/worker/IntrinsicMapTaskExecutor.java
@@ -20,8 +20,8 @@ package org.apache.beam.runners.dataflow.worker;
 import com.google.api.services.dataflow.model.CounterUpdate;
 import java.util.Collections;
 import java.util.List;
+import org.apache.beam.runners.core.metrics.ExecutionStateTracker;
 import org.apache.beam.runners.dataflow.worker.counters.CounterSet;
-import org.apache.beam.runners.dataflow.worker.util.common.worker.ExecutionStateTracker;
 import org.apache.beam.runners.dataflow.worker.util.common.worker.Operation;
 
 /**
